feat(ui): Add top bar nav.
