--- web/quickplan/src/AntdCustomComponents/DatePicker.tsx
@@ -1,6 +1,0 @@
-import {Dayjs} from 'dayjs';
-import dayjsGenerateConfig from 'rc-picker/lib/generate/dayjs';
-import generatePicker from 'antd/es/date-picker/generatePicker';
-import 'antd/es/date-picker/style/index';
-
-export const DatePicker = generatePicker<Dayjs>(dayjsGenerateConfig);
