--- web/quickplan/src/AntdCustomComponents/README.md
@@ -1,5 +1,0 @@
-# AntdCustomComponents
-
-This directory exists because Antd uses Moment.js by default, and we want to use `dayjs`.
-
-You can see the instructions implemented by the files in this directory [here](https://ant.design/docs/react/replace-moment).
