+++ web/quickplan/src/components/AntdCustomComponents/DatePicker.tsx
@@ -1,0 +1,6 @@
+import {Dayjs} from 'dayjs';
+import dayjsGenerateConfig from 'rc-picker/lib/generate/dayjs';
+import generatePicker from 'antd/es/date-picker/generatePicker';
+import 'antd/es/date-picker/style/index';
+
+export const DatePicker = generatePicker<Dayjs>(dayjsGenerateConfig);
