+++ web/quickplan/src/components/Layout/TopNav.tsx
@@ -1,0 +1,15 @@
+import React, {FC} from "react";
+import {Menu} from "antd";
+
+export const TopNav: FC = () => {
+    const items = [
+        {key: 'export-graphiz', label: 'Export Graphviz'}
+    ]
+    return (
+        <Menu mode="horizontal"
+              items={items}
+              style={{background: '#fff', position: 'relative'}}
+        />
+    )
+
+}
